Add files via upload
Sorry, I'm having trouble taking a picture of the diagram I've been working to show how I'm setting up the database. (It's a little funky and inconsistent). I've been doing it in Dia, which is free & open source so the diagram here can be opened in Dia. Tomorrow night I'll finish that up and maybe do a quick simple one that I can take a picture of (I hope) to show you what I have in mind. I think we'll be meeting Wed. and I can start coding the DB Thurs. If we decide to do it differently, I can always change it.